Start towards removing map_color_rgb.
There is a an HWColorMap parameter, generated by param_HWComp that
describes the palette for a given device. This relies on calling
map_color_rgb on all the possible color values. Thus this generates
the palette from the device rgb values, not color correct ones.

We hope to remove the (long deprecated) map_color_rgb device proc,
and replace it with calls to decode_color. The wrinkle here is
that decode_color decodes to the native device colorspace, which
may not be rgb.

Accordingly, we introduce a gx_map_rgb_color helper function that
performs a decode_color, and then (depending on the number of
components) converts it (in a non-color correct manner) to rgb if
required.

This is imperfect, but should be no worse than what we have now -
indeed, potentially better, because frequently map_color_rgb is
set to be decode_color, so this would give the wrong results for
cmyk or greyscale devices.

The bmp devices do a similar operation when generating the palette
in their headers; use the same routine for that too.